refactor: create open or active tab method
