+++ packages/shared-base-ui/src/bom/index.ts
@@ -1,1 +1,2 @@
 export { openWindow } from './open-window.js'
+export { openOrActiveTab } from './open-tab.js'
